Fix fragment exception impacts queued fragments (#9041)
## Describe your changes

In the current state, an exception in a fragment will cause queued
fragments to stop from running. But instead we want queued fragments to
still run even if a previous fragment raised an exception (even if the
error happens in a full app run; the error will be displayed but the app
will continue to run).

To do so, we add except blocks into the loop and continue running the
other fragments in the queue.

Also, we want to stop app execution when a fragment raises an error
during a full app run. This is why the `exec_func_with_error_handling`
call is removed in `fragment.py`.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.